requirements.txt
langchain
streamlit
pymongo
langgraph
python-dotenv
langchain-groq
langchain-pinecone
langchain-core
langchain-community
pinecone
lark
transformers
langsmith
langchain-openai